TIP-706: move price filter
